Gameplan
I commented out a bunch of code as well as added a few functions to break down my idea. I'm going to try to work on one function per day so I can check off each one as I go. If I need to go back and edit a function that's fine, as there's a chance my prior code won't meet the requirements of the future.