Error fix, tuning, setup, and theory
I have cleaned up all errors present at the current time, reformatted the code to be easier to read, fixed some bad function calculations, as well as added the correct sensor and motor functions. As far as theory goes, I've figured out how I plan to calculate what constitutes a hole, as well as how to identify a hole.

I've also realized that if we have a coordinate grid, we may be able to print those coordinates to a file and read off them for debugging purposes. Chances are we could be able to draw a map of the area from the coordinates for easy visualization of what the robot is seeing/calculating.