Completed fit test?
I think I've covered my bases for the fit test function. I think the next step is creating decision nodes. Once I solve that issue I can easily make a function to return to a given node. Then I can write the driving function and start to debug (aka the scary part). I can test with either a given set of points (aka simulate scanned coordinates and skip the scan feature) or I can change the sensor within the code to the one use in the VEX VR Simulator and test it on the dynamic maze. Sure, it may not be accurate to mazes we may be given in the real world, but it should work regardless of the data fed to it.