Revisions and the Start of Version 2
I've added a little bit to version one, and I started a second version that will be better for this use case. I think if I kept writing version one it could get to a functional point, but I think if I focus the scope of the project to what will be needed, we'll have a better solution.